Move SPARQL queries into the Report that use them
Move towards each Report file being self-contained, so that adding a
new type of report largely involves only adding a single file.
